% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbgtest_feols.R
\name{pbgtest_feols}
\alias{pbgtest_feols}
\title{Pbgtest_feols: Performs Breusch-Godfrey/Wooldridge test for serial
correlation in panel data using Fixed Effects Ordinary Least Squares (FEOLS)
estimation.
Args:
This function calculates the Breusch-Godfrey/Wooldridge test statistic to
assess serial correlation in panel data models. It computes the test statistic
and corresponding p-value, indicating whether there is evidence of serial
correlation in the idiosyncratic errors. The test is conducted under the
assumption of Fixed Effects Ordinary Least Squares (FEOLS) estimation.}
\usage{
pbgtest_feols(model)
}
\arguments{
\item{model}{}
}
\value{
the results of the pdg test on fixest objects
}
\description{
Pbgtest_feols: Performs Breusch-Godfrey/Wooldridge test for serial
correlation in panel data using Fixed Effects Ordinary Least Squares (FEOLS)
estimation.
Args:
This function calculates the Breusch-Godfrey/Wooldridge test statistic to
assess serial correlation in panel data models. It computes the test statistic
and corresponding p-value, indicating whether there is evidence of serial
correlation in the idiosyncratic errors. The test is conducted under the
assumption of Fixed Effects Ordinary Least Squares (FEOLS) estimation.
}
\examples{
pbgtest_feols(feols(y ~ x | i + t, data = pdata, panel.id = c('i', 't')))
}
